Refactor JobStatusRetrievalStrategy unit tests (#280)
